+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - ItemViewModel.bind `defaultValue` parameter
 - Node builders inside of `MenuItem` will automatically assign the node to the `graphic` property of the menu item
 - The App class (main application entrypoint) no longer requires a primary view parameter, in case you want to show a tray icon or determinine what view to show some other way
+- Renamed tableview column builder for readonly non-observable properties to `readonlyColumn` (https://github.com/edvin/tornadofx/issues/599)
 
 ### Additions
 - TextInputControl.filterInput allows you to discriminate what kind of input should be accepted for a text input control